# specs/029-llm-friendly-bot-operations/contracts/api.yaml
openapi: 3.0.0
info:
  title: Trading Bot LLM Operations API
  description: |
    REST API for LLM-friendly bot operations, monitoring, and management.

    This API provides structured interfaces for LLMs to:
    - Query bot state and performance
    - Monitor health and errors
    - Execute maintenance workflows
    - Manage configuration
    - Stream real-time updates
  version: 1.0.0
  contact:
    name: Bot Operations Team
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api.bot.example.com
    description: Production
security:
  - ApiKeyAuth: []
paths:
...
  /api/v1/workflows:
    get:
      summary: List available workflows
      description: Returns all defined maintenance workflows
      tags: [Workflows]
      responses:
        '200':
          description: Workflows retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Workflow'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/workflows/{workflow_id}/execute:
    post:
      summary: Execute maintenance workflow
      description: Starts workflow execution with progress tracking
      tags: [Workflows]
      parameters:
        - in: path
          name: workflow_id
          required: true
          schema:
            type: string
          description: Workflow identifier (e.g., "restart-bot", "update-targets")
      responses:
        '202':
          description: Workflow execution started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowExecution'
        '400':
          description: Workflow validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SemanticError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Workflow not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SemanticError'
  /api/v1/workflows/{workflow_id}/status:
    get:
      summary: Get workflow execution status
      description: Returns current status and progress of workflow execution
      tags: [Workflows]
      parameters:
        - in: path
          name: workflow_id
          required: true
          schema:
            type: string
          description: Workflow identifier
      responses:
        '200':
          description: Workflow status retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowExecution'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Workflow execution not found
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
  schemas:
...
    Workflow:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        steps:
          type: array
          items:
            $ref: '#/components/schemas/WorkflowStep'
    WorkflowStep:
      type: object
      properties:
        id:
          type: string
        description:
          type: string
        action_type:
          type: string
          enum: [command, api_call, validation, wait]
    WorkflowExecution:
      type: object
      properties:
        workflow_id:
          type: string
        status:
          type: string
          enum: [NOT_STARTED, IN_PROGRESS, COMPLETED, FAILED]
        current_step:
          type: integer
        total_steps:
          type: integer
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
        error:
          $ref: '#/components/schemas/SemanticError'
          nullable: true
    SemanticError:
      type: object
      required: [error_code, error_type, message, cause, impact, remediation, context, timestamp, severity]
      properties:
        error_code:
          type: string
          description: Machine-readable error code (e.g., BOT_001, API_429)
        error_type:
          type: string
          description: Error category
        message:
          type: string
          description: Human-readable error description
        cause:
          type: string
          description: Root cause explanation
        impact:
          type: string
          description: Consequences of this error
        remediation:
          type: string
          description: Specific steps to resolve
        context:
          type: object
          additionalProperties: true
          description: Relevant identifiers for correlation
        timestamp:
          type: string
          format: date-time
        severity:
          type: string
          enum: [low, medium, high, critical]
...
  responses:
    Unauthorized:
      description: Unauthorized - invalid or missing API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SemanticError'
          example:
            error_code: "AUTH_001"
            error_type: "AuthenticationError"
            message: "Invalid or missing API key"
            cause: "X-API-Key header not provided or token invalid"
            impact: "API access denied"
            remediation: "Provide valid API key in X-API-Key header"
            context: {}
            timestamp: "2025-10-24T14:32:00Z"
            severity: "high"
...